requirements.txt
# AI Tomato Sorter - Requirements
# Core ML and Computer Vision
ultralytics>=8.0.0
opencv-python-headless>=4.7.0
numpy>=1.23.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0

# Model Export and Optimization
onnx>=1.14.0
onnxruntime>=1.15.0
tensorflow>=2.13.0
# tflite-runtime>=2.11.0  # Install platform-specific wheel on Pi

# Data Processing and Augmentation
albumentations>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Raspberry Pi Specific
RPi.GPIO>=0.7.1
pyserial>=3.5
flask>=2.3.0
flask-socketio>=5.3.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Development and Testing
pytest>=7.4.0
jupyter>=1.0.0
ipywidgets>=8.0.0
